Corrected env.travis to use a sqlite database during tests + corrected migrations to be compatible with sqlite databases

+++ database/migrations/2017_11_07_135119_doc_user_action_column_wrong_type.php
@@ -13,13 +13,15 @@ class DocUserActionColumnWrongType extends Migration
      */
     public function up()
     {
-        Schema::table('document_user', function (Blueprint $table) {
-            $table->dropColumn(["action"]);
-        });
+        if (Schema::hasColumn('document_user', 'action')) {
+            Schema::table('document_user', function (Blueprint $table) {
+                $table->dropColumn(["action"]);
+            });
+        }
 
 
         Schema::table('document_user', function (Blueprint $table) {
-            $table->string('action');
+            $table->string('action')->nullable();
         });
     }
